pyproject.toml
[project]
name = "app-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.124.4",
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.7",
    "langchain-openai>=1.1.3",
    "langgraph>=0.6.3",
    "pip>=25.2",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.1.1",
    "tortoise-orm>=0.25.2",
    "uvicorn[standard]>=0.38.0",
]

[tool.aerich]
tortoise_orm = "agent_v1.api.db.config.tortoise_config"
location = "./migrations"
src_folder = "./."
